+++ lib/src/main/java/com/example/lib/controller/LibController.java
@@ -2,6 +2,7 @@ package com.example.lib.controller;
 
 import com.example.lib.dto.BookStatusDto;
 import com.example.lib.service.LibService;
+import com.example.lib.jwt.JwtProvider;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
@@ -16,6 +17,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.DeleteMapping;
 
+import jakarta.servlet.http.HttpServletRequest;
+
 import java.util.List;
 
 @RestController
@@ -24,6 +27,7 @@ import java.util.List;
 public class LibController {
 
     private final LibService libraryService;
+    private final JwtProvider jwtProvider;
 
     @Operation(summary = "Получить список свободных книг", description = "Возвращает список книг, доступных для взятия")
     @ApiResponses(value = {
@@ -31,19 +35,32 @@ public class LibController {
             @ApiResponse(responseCode = "500", description = "Внутренняя ошибка сервера")
     })
     @GetMapping("/books/available")
-    public ResponseEntity<List<BookStatusDto>> getAvailableBooks() {
+    public ResponseEntity<List<BookStatusDto>> getAvailableBooks(HttpServletRequest request) {
+        String token = jwtProvider.resolveToken(request);
+        if (token == null || !jwtProvider.validateToken(token)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
+
+
         List<BookStatusDto> availableBooks = libraryService.getAvailableBooks();
         return new ResponseEntity<>(availableBooks, HttpStatus.OK);
     }
-    //используется только чтобы получить запрос на удаление из bookservice
+
     @Operation(summary = "Удалить книгу", description = "Удаляет книгу по указанному ID")
     @DeleteMapping("/book/{bookId}")
-    public ResponseEntity<Void> deleteBookStatus(@PathVariable Long bookId) {
+    public ResponseEntity<Void> deleteBookStatus(@PathVariable Long bookId, HttpServletRequest request) {
+        String token = jwtProvider.resolveToken(request);
+
+        // Проверка, что токен валиден и пользователь имеет роль администратора
+        if (token == null || !jwtProvider.validateToken(token) ||
+                !jwtProvider.getRoleFromToken(token).contains("ROLE_ADMIN")) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN); // Если проверка не пройдена, вернется 403 Forbidden
+        }
+
+        // Если проверка пройдена, удалить книгу
         libraryService.deleteBookStatus(bookId);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
-
-    //используется только чтобы получить запрос на добавление книги из bookservice
 
     @Operation(summary = "Добавление книги", description = "Добавляет книгу в систему на основе bookId")
     @ApiResponses(value = {
@@ -52,10 +69,18 @@ public class LibController {
             @ApiResponse(responseCode = "500", description = "Внутренняя ошибка сервера")
     })
     @PostMapping("/book")
-    public ResponseEntity<BookStatusDto> addBook(@RequestParam Long bookId) {
+    public ResponseEntity<BookStatusDto> addBook(@RequestParam Long bookId, HttpServletRequest request) {
+        String token = jwtProvider.resolveToken(request);
+
+        // Проверка, что токен валиден и пользователь имеет роль администратора
+        if (token == null || !jwtProvider.validateToken(token) ||
+                !jwtProvider.getRoleFromToken(token).contains("ROLE_ADMIN")) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN); // Если проверка не пройдена, вернется 403 Forbidden
+        }
         BookStatusDto bookStatusDto = libraryService.addBook(bookId);
         return new ResponseEntity<>(bookStatusDto, HttpStatus.CREATED);
     }
+
 
     @Operation(summary = "Взять книгу", description = "Помечает книгу как взятую и устанавливает дату возврата через 14 дней")
     @ApiResponses(value = {
@@ -64,7 +89,14 @@ public class LibController {
             @ApiResponse(responseCode = "500", description = "Внутренняя ошибка сервера")
     })
     @PostMapping("/book/borrow")
-    public ResponseEntity<BookStatusDto> borrowBook(@RequestParam Long bookId) {
+    public ResponseEntity<BookStatusDto> borrowBook(@RequestParam Long bookId,HttpServletRequest request) {
+        String token = jwtProvider.resolveToken(request);
+
+        // Проверка, что токен валиден и пользователь имеет роль администратора
+        if (token == null || !jwtProvider.validateToken(token) ||
+                !jwtProvider.getRoleFromToken(token).contains("ROLE_ADMIN")) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN); // Если проверка не пройдена, вернется 403 Forbidden
+        }
         BookStatusDto bookStatusDto = libraryService.borrowBook(bookId);
         return new ResponseEntity<>(bookStatusDto, HttpStatus.OK);
     }
@@ -76,7 +108,14 @@ public class LibController {
             @ApiResponse(responseCode = "500", description = "Внутренняя ошибка сервера")
     })
     @PostMapping("/book/return")
-    public ResponseEntity<BookStatusDto> returnBook(@RequestParam Long bookId) {
+    public ResponseEntity<BookStatusDto> returnBook(@RequestParam Long bookId,HttpServletRequest request) {
+        String token = jwtProvider.resolveToken(request);
+
+        // Проверка, что токен валиден и пользователь имеет роль администратора
+        if (token == null || !jwtProvider.validateToken(token) ||
+                !jwtProvider.getRoleFromToken(token).contains("ROLE_ADMIN")) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN); // Если проверка не пройдена, вернется 403 Forbidden
+        }
         BookStatusDto bookStatusDto = libraryService.returnBook(bookId);
         return new ResponseEntity<>(bookStatusDto, HttpStatus.OK);
     }
@@ -88,7 +127,13 @@ public class LibController {
             @ApiResponse(responseCode = "500", description = "Внутренняя ошибка сервера")
     })
     @GetMapping("/book/status")
-    public ResponseEntity<BookStatusDto> getBookStatus(@RequestParam Long bookId) {
+    public ResponseEntity<BookStatusDto> getBookStatus(@RequestParam Long bookId,HttpServletRequest request) {
+        String token = jwtProvider.resolveToken(request);
+
+        // Проверка, что токен валиден и пользователь имеет роль администратора
+        if (token == null || !jwtProvider.validateToken(token) ) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN); // Если проверка не пройдена, вернется 403 Forbidden
+        }
         BookStatusDto bookStatusDto = libraryService.getBookStatus(bookId);
         return new ResponseEntity<>(bookStatusDto, HttpStatus.OK);
     }
